+++ src/masonite/notifications/NotificationFacade.py
@@ -1,0 +1,51 @@
+"""Base Notification facade."""
+from abc import ABC, abstractmethod
+
+
+class NotificationFacade(ABC):
+    """Base Notification facade."""
+
+    def __init__(self, *args, **kwargs):
+        self.id = None
+        self._run = True
+        self._fail_silently = False
+
+    def broadcast_on(self):
+        """Get the channels the event should broadcast on."""
+        return []
+
+    @abstractmethod
+    def via(self, notifiable):
+        """Defines the notification's delivery channels."""
+        return []
+
+    @property
+    def should_send(self):
+        return self._run
+
+    @property
+    def ignore_errors(self):
+        return self._fail_silently
+
+    def dry(self):
+        """Sets whether the notification should be sent or not.
+
+        Returns:
+            self
+        """
+        self._run = False
+        return self
+
+    def fail_silently(self):
+        """Sets whether the notification can fail silently (without raising exceptions).
+
+        Returns:
+            self
+        """
+        self._fail_silently = True
+        return self
+
+    @classmethod
+    def notification_type(cls):
+        """Get notification type defined with class name."""
+        return cls.__name__
