+++ src/masonite/notifications/__init__.py
@@ -1,6 +1,6 @@
 from .Notifiable import Notifiable
-from .Notification import Notification
+from .NotificationFacade import NotificationFacade
 from .models import DatabaseNotification
-from .Notify import Notify
+from .Notification import Notification
 from .NotificationContract import NotificationContract
 from .providers import NotificationProvider
